+++ src/boot/polyfills/string.prototype.normalize.js
